Fix a bug in computing La_SL
Fix a bug that La_SL may be zero or close to zero when the wind is
very small. Fixed by checking if both the surface Stokes drift and
the surface friction velocity are greater than some small value
(here 1e-4 m/s). If not, set La_SL to the default big number (1e8).